請兩個方法，請老師幫我看看
