+++ src/main/java/com/github/hcsp/multithread/MultiThreadWordCount3.java
@@ -4,40 +4,55 @@ import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.locks.Condition;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
-public class MultiThreadWordCount2 {
-    // 使用threadNum个线程，并发统计文件中各单词的数量 Lock/Condition
-    public static Map<String, Integer> count(int threadNum, List<File> files) throws IOException {
-        final Lock lock = new ReentrantLock();
-        Condition condition = lock.newCondition();
+public class MultiThreadWordCount3 {
+    // 使用threadNum个线程，并发统计文件中各单词的数量 CountDownLatch
+
+    public static void main(String[] args) throws IOException, InterruptedException {
+        List<File> files = Arrays.asList(new File("c:/users/sunp/git/multithread-word-count/target/test.txt"));
+        System.out.println(count(10, files));
+    }
+
+    public static Map<String, Integer> count(int threadNum, List<File> files) throws IOException, InterruptedException {
         Map<String, Integer> result = new HashMap<>();
         for (File file : files) {
+            List<Map<String, Integer>> middleList = new ArrayList<>();
             BufferedReader bReader = Files.newBufferedReader(file.toPath());
+            ExecutorService executorService = Executors.newFixedThreadPool(threadNum);
+            CountDownLatch latch = new CountDownLatch(threadNum);
             for (int i = 0; i < threadNum; i++) {
-                new Thread(() -> {
-                    lock.lock();
+                Map<String, Integer> middleResult = new HashMap<>();
+                Runnable runnable = () -> {
                     try {
-                        String line = bReader.readLine();
-                        while (null != line) {
+                        String line;
+                        while (null != (line = bReader.readLine())) {
                             String[] strings = line.split(" ");
-                            condition.await();
                             for (String word : strings) {
-                                result.put(word, result.getOrDefault(word, 0) + 1);
+                                middleResult.put(word, middleResult.getOrDefault(word, 0) + 1);
                             }
-                            condition.signal();
                         }
-                    } catch (IOException | InterruptedException e) {
+                        middleList.add(middleResult);
+                    } catch (IOException e) {
                         throw new RuntimeException(e);
                     } finally {
-                        lock.unlock();
+                        latch.countDown();
                     }
-                }).start();
+                };
+                executorService.execute(runnable);
+            }
+            latch.await();
+            for (Map<String, Integer> stringIntegerMap : middleList) {
+                stringIntegerMap.forEach((key, value) -> {
+                    result.put(key, result.getOrDefault(key, 0) + value);
+                });
             }
         }
         return result;
