+++ src/main/java/com/github/hcsp/multithread/MultiThreadWordCount4.java
@@ -5,49 +5,55 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Files;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
+import java.util.concurrent.*;
 
-public class MultiThreadWordCount3 {
-    // 使用threadNum个线程，并发统计文件中各单词的数量 CountDownLatch
-    public static Map<String, Integer> count(int threadNum, List<File> files) throws IOException, InterruptedException {
+public class MultiThreadWordCount4 {
+
+    // 使用threadNum个线程，并发统计文件中各单词的数量 feature and thread pool
+
+    public static void main(String[] args) throws IOException, InterruptedException, ExecutionException {
+        List<File> files = Arrays.asList(new File("c:/users/sunp/git/multithread-word-count/target/test.txt"),
+                new File("c:/users/sunp/git/multithread-word-count/target/test2.txt"));
+        System.out.println(count(10, files));
+    }
+
+    public static Map<String, Integer> count(int threadNum, List<File> files)
+            throws IOException, InterruptedException, ExecutionException {
         Map<String, Integer> result = new HashMap<>();
-        List<Map<String, Integer>> middleList = new ArrayList<>();
         for (File file : files) {
-            BufferedReader bReader = Files.newBufferedReader(file.toPath());
+            List<Map<String, Integer>> middleList = new ArrayList<>();
             ExecutorService executorService = Executors.newFixedThreadPool(threadNum);
-            CountDownLatch latch = new CountDownLatch(threadNum);
+            BufferedReader bReader = Files.newBufferedReader(file.toPath());
             for (int i = 0; i < threadNum; i++) {
-                Map<String, Integer> middleResult = new HashMap<>();
-                Runnable runnable = () -> {
+            Future<Map<String, Integer>> future = executorService.submit(new Callable<Map<String, Integer>>() {
+                @Override
+                public Map<String, Integer> call() throws Exception {
                     try {
+                        Map<String, Integer> middleResult = new HashMap<>();
                         String line;
-                        while ((line = bReader.readLine()) != null) {
+                        while (null != (line = bReader.readLine())) {
                             String[] strings = line.split(" ");
                             for (String word : strings) {
                                 middleResult.put(word, middleResult.getOrDefault(word, 0) + 1);
                             }
                         }
-                        middleList.add(middleResult);
+                        return middleResult;
                     } catch (IOException e) {
                         throw new RuntimeException(e);
-                    } finally {
-                        latch.countDown();
                     }
-                };
-                executorService.execute(runnable);
+                }
+            });
+            middleList.add(future.get());
             }
-            latch.await();
             for (Map<String, Integer> stringIntegerMap : middleList) {
                 stringIntegerMap.forEach((key, value) -> {
                     result.put(key, result.getOrDefault(key, 0) + value);
                 });
             }
-//            bReader.close();
         }
         return result;
     }
